BUG: Fixed mismatches of data type between FITS columns and C++ variables
git-svn-id: https://elves.srl.caltech.edu/svn/nusim/trunk@294 f61b1093-7add-4ab6-ac1b-eb05caf01ebb

+++ src/modules/inc/NModuleInterfaceEventSaverLevel2Fits.h
@@ -89,9 +89,9 @@ class NModuleInterfaceEventSaverLevel2Fits
   vector<double> m_Dec;
   vector<double> m_Energy;
   vector<double> m_Time;
-  vector<double> m_Origin;
-  vector<double> m_PI;
-  vector<double> m_Grade;
+  vector<int>    m_Origin;
+  vector<int>    m_PI;
+  vector<int>    m_Grade;
 
 
 #ifdef ___CINT___
